remove dead styles in sidemenu move layout stylesheets to styles dir move layout components into root layout dir delete empty layout dirs
