+++ src/layout/index.ts
@@ -1,4 +1,4 @@
-export * from './Footer/Footer'
-export * from './Nav/Nav'
-export * from './Nav/SideMenu/SideMenu'
+export * from './Footer'
+export * from './Nav'
+export * from './SideMenu'
 export * from './Layout'
